pyproject.toml
[project]
name = "delta-stream-experiment"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = ["click>=8.1.8", "delta-spark>=3.3.1", "pyspark>=3.5.5"]

[dependency-groups]
dev = ["mypy>=1.15.0", "ruff>=0.11.7"]

[tool.ruff.lint]
extend-select = ["I"]
